+++ workspace/ims-webapp/src/main/java/th/go/excise/ims/ta/persistence/entity/TaPaperPrRedTax.java
@@ -26,7 +26,7 @@ public class TaPaperPrRedTax extends BaseEntity {
 	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "TA_PAPER_PR_RED_TAX_GEN")
 	@SequenceGenerator(name = "TA_PAPER_PR_RED_TAX_GEN", sequenceName = "TA_PAPER_PR_RED_TAX_SEQ", allocationSize = 1)
 	@Column(name = "PAPER_PR_RED_TAX_ID")
-	private BigDecimal paperPrRedTaxId;
+	private Long paperPrRedTaxId;
 	@Column(name = "PLAN_NUMBER")
 	private String planNumber;
 	@Column(name = "PAPER_NUMBER")
@@ -64,11 +64,11 @@ public class TaPaperPrRedTax extends BaseEntity {
 	@Column(name = "DIFF_TAX_REDUCE_AMT")
 	private BigDecimal diffTaxReduceAmt;
 
-	public BigDecimal getPaperPrRedTaxId() {
+	public Long getPaperPrRedTaxId() {
 		return paperPrRedTaxId;
 	}
 
-	public void setPaperPrRedTaxId(BigDecimal paperPrRedTaxId) {
+	public void setPaperPrRedTaxId(Long paperPrRedTaxId) {
 		this.paperPrRedTaxId = paperPrRedTaxId;
 	}
