+++ workspace/ims-webapp/src/main/java/th/go/excise/ims/ta/persistence/repository/TaPaperPrOutFrnGoodsRepository.java
@@ -1,10 +1,9 @@
 
 package th.go.excise.ims.ta.persistence.repository;
 
-import java.math.BigDecimal;
 import th.co.baiwa.buckwaframework.common.persistence.repository.CommonJpaCrudRepository;
 import th.go.excise.ims.ta.persistence.entity.TaPaperPrOutFrnGoods;
 
-public interface TaPaperPrOutFrnGoodsRepository extends CommonJpaCrudRepository<TaPaperPrOutFrnGoods, BigDecimal> {
+public interface TaPaperPrOutFrnGoodsRepository extends CommonJpaCrudRepository<TaPaperPrOutFrnGoods, Long> {
 
 }
