+++ workspace/ims-webapp/src/main/java/th/go/excise/ims/ta/persistence/repository/TaPaperPrOutGoodsRepository.java
@@ -1,10 +1,9 @@
 
 package th.go.excise.ims.ta.persistence.repository;
 
-import java.math.BigDecimal;
 import th.co.baiwa.buckwaframework.common.persistence.repository.CommonJpaCrudRepository;
 import th.go.excise.ims.ta.persistence.entity.TaPaperPrOutGoods;
 
-public interface TaPaperPrOutGoodsRepository extends CommonJpaCrudRepository<TaPaperPrOutGoods, BigDecimal> {
+public interface TaPaperPrOutGoodsRepository extends CommonJpaCrudRepository<TaPaperPrOutGoods, Long> {
 
 }
